djangocms_reversion2/requirements.txt
django>=1.8.17,<1.9
django-cms==3.4.2
diff-match-patch==20110725.1
django-sekizai==0.9.0
xlsxwriter
djangocms-text-ckeditor
djangocms-text-ckeditor==3.5.0
